fix(localresolver): temp fix for inconsistent keyRef formats
The `id` for keys in the verificationMethod section of a did doc returned by validateEvents is a
KERI prefix (i.e. D...), while keys in the VKP have their base64 form as their `id` (i.e. no D
prefix and an extra padding char =). This fixes the issue for now, until it can be sorted in the
appropriate dependency.