Initial credentialBuilder implementation
A new Verifiable Credential class is defined (compliant to v1.1 of the specification).
A set of simple tests is added to ensure the new class can be instantiated from valid JSON, with all fields set as expected.
Added an initial definition of the ChainedProof2021 proof type. Further work is needed to polish the API and make certain properties (i.e. chainSignatureSuite) easily configurable.